working on mvc model for the user
