+++ package.json
@@ -40,6 +40,7 @@
     "dotenv": "^16.0.3",
     "eslint": "^8.37.0",
     "express": "^4.18.2",
+    "express-validator": "^7.0.1",
     "jsonwebtoken": "^9.0.0",
     "mocha": "^10.2.0",
     "nodemon": "^2.0.22",
