+++ prisma/schema.prisma
@@ -12,32 +12,25 @@ model Address {
   state     String @db.VarChar(255)
   city      String @db.VarChar(255)
   street    String @db.VarChar(255)
-  user_Id   Int
-  user      User   @relation(fields: [user_Id], references: [Id])
+  user      User[] // two users can have the same address
 }
 
 model User {
   userId       String        @unique @default(uuid())
   Id           Int           @id @default(autoincrement())
+  first_name   String        @db.VarChar(255)
+  last_name    String        @db.VarChar(255)
   username     String        @unique(map: "username") @db.VarChar(255)
+  phone_number String        @db.VarChar(25)
   email        String        @unique(map: "email") @db.VarChar(255)
   created_at   DateTime      @default(now())
   updated_at   DateTime      @updatedAt
   password     String        @db.VarChar(255)
+  addressId    Int? //each user has one address
   RefreshToken RefreshToken?
-  userprofile  Userprofile?
-  address      Address[]
-}
+  address      Address?      @relation(fields: [addressId], references: [addressId])
 
-model Userprofile {
-  userprofileId Int    @id @default(autoincrement())
-  first_name    String @db.VarChar(255)
-  last_name     String @db.VarChar(255)
-  phone_number  String @db.VarChar(25)
-  user          User   @relation(fields: [user_Id], references: [Id])
-  user_Id       Int    @unique
-
-  @@index([user_Id], map: "userprofile_id_fkey")
+  @@unique([first_name, last_name]) // they can have the same name but not the same email
 }
 
 model RefreshToken {
