+++ src/middlewares/validateregistration.ts
@@ -1,32 +1,19 @@
-import { Response, NextFunction} from 'express';
-import { findByUsernameorEmail} from '../db/users';
+import { Request, Response, NextFunction } from 'express';
+// check handles req param, validationresult handles errors
+import { body, validationResult} from 'express-validator';
 
-export const validateRegistration = async (req: any, res: Response, next:NextFunction) => {
-    const {email, password, username} = req.body;
+export const validateRegistration = async (req: Request, res: Response, next: NextFunction) => {
 
-    if (!email) {
-        return res.status(400).json({message: 'email is required'});
-    }
-    if (!password) {
-        return res.status(400).json({message: 'password is required'});
-    }
-    if (!username) {
-        return res.status(400).json({message: 'username is required'});
-    }
+    const { user } = req.body
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 
-    const user = await findByUsernameorEmail(username, email);
-    if (user) {
-       let error = 'user already exists';
-       if (user.email !== email) {
-              error = 'username already exists';
-              return res.json({message: error});
-         }
-    }
-        req.user = {
-            email,
-            username,
-            password
-        }
+   
 
+
+
+
+   
     next();
 }
+
+
