+++ src/server.ts
@@ -16,6 +16,7 @@ app.use('/api/v1', Authrouter);
 app.use('api/v1/register', Authrouter);
 app.use('api/v1/login', Authrouter);
 app.use('api/v1/update/:userId', Authrouter);
+app.use('api/v1/update/address/:user_Id', Authrouter);
 app.listen(3000, () => {
     console.log(`Server running on port ${port}`);
     });
